57: Remove coming soon page
